+++ src/main/java/com/example/demo/controller/ViewController.java
@@ -9,7 +9,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.bind.annotation.RequestMethod;
-import twitter4j.*;
 
 @RestController
 @RequestMapping(path = "/")
@@ -31,45 +30,5 @@ public class ViewController {
         List<Map<String, Object>> list;
         list = jdbcTemplate.queryForList("select * from users where id = ?", id);
         return list.toString();
-    }
-
-    @RequestMapping(path = "/search_keyword", method = RequestMethod.GET)
-    public String search_keyowrd(@RequestParam(name = "searchword", defaultValue = "チンチン") String searchword) throws TwitterException {
-        // 初期化
-        Twitter twitter = new TwitterFactory().getInstance();
-        Query query = new Query(searchword);
-
-        QueryResult result = twitter.search(query);
-
-        // for debug. show tweet results on CLI
-        for (Status status : result.getTweets()) {
-            System.out.println("@" + status.getUser().getScreenName() + ":" + status.getText());
-        }
-
-        return "a";
-    }
-
-    @RequestMapping(path = "/search_user", method = RequestMethod.GET)
-    public String search_user(@RequestParam(name = "username", defaultValue = "CNN") String username,
-                              @RequestParam(name = "tweet_id", defaultValue = "1000000") long tweet_id
-    ) throws TwitterException {
-        // 初期化
-        Twitter twitter = new TwitterFactory().getInstance();
-        Query query = new Query();
-
-        //ユーザーネームで指定.sinceId以降のtweetから15個取ってくる.最新のツイートから時系列順に取ってくるっぽい。
-        query.setQuery("from:" + username);
-        query.setSinceId(tweet_id);
-
-        QueryResult result = twitter.search(query);
-
-        //for debug on CLI.
-        for (Status status : result.getTweets()) {
-            System.out.println("@" + status.getUser().getScreenName() + ":" + status.getText());
-        }
-
-        //TODO: ここで取ってきたLatestTweetIdを保存しておいて、次回からそのidのツイート以降のツイートを拾ってくればOK
-
-        return "a";
     }
 }
